feat: refactor SDK
