+++ Cargo.toml
@@ -33,9 +33,7 @@ solana-account-decoder = "~2.0"
 solana-cli-config = "~2.0"
 yellowstone-grpc-client = "3.1.1"
 yellowstone-grpc-proto = "3.1.1"
-risc0-zkvm = { version = "1.2.1", features = [
-  "prove",
-], default-features = false }
+risc0-zkvm = { version = "1.2.1", default-features = false, features = ["prove"] }
 risc0-core = { version = "1.2.1" }
 risc0-zkp = { version = "1.2.1" }
 risc0-groth16 = { version = "1.2.1" }
@@ -47,6 +45,7 @@ bonsol-interface = { version = "0.4.5", path = "./onchain/interface" }
 bonsol-schema = { version = "0.4.5", path = "./schemas" }
 bonsol-sdk = { version = "0.4.5", path = "./sdk" }
 bonsol-prover = { version = "0.4.5", path = "./prover" }
+tokio = { version = "1.44.1" }
 
 [patch.crates-io.curve25519-dalek]
 git = "https://github.com/anza-xyz/curve25519-dalek.git"
