+++ local-zk-program-server/Cargo.toml
@@ -5,7 +5,7 @@ version = { workspace = true }
 
 [dependencies]
 actix-web = "4.10.2"
-tokio = {version = "1.44.1", features = ["full"]}
+tokio = { workspace = true, features = ["full"] }
 
 [lints]
 workspace = true
