+++ node/Cargo.toml
@@ -67,7 +67,7 @@ solana-transaction-status = { workspace = true }
 strum_macros = "0.26.4"
 tempfile = "3.10.1"
 thiserror = { workspace = true }
-tokio = { version = "1.36.0", features = ["full"] }
+tokio = { workspace = true, features = ["full"] }
 tokio-util = { version = "0.7.10", features = ["codec"] }
 tracing = "0.1.40"
 tracing-subscriber = { version = "0.3.18", features = [
