+++ sdk/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "bonsol-sdk"
-version.workspace = true
+version = { workspace = true }
 description = "SDK for interacting with Bonsol protocol"
 authors = ["anagram build team"]
 repository = "https://github.com/bonsolcollective/bonsol"
@@ -11,23 +11,24 @@ edition = "2021"
 anyhow = "1.0.86"
 async-trait = "0.1.80"
 bincode = "1.3.3"
-bonsol-interface.workspace = true
-bonsol-schema.workspace = true
+bonsol-interface = { workspace = true }
+bonsol-schema = { workspace = true }
 bytes = "1.5.0"
 flatbuffers = "24.3.25"
 futures-util = "0.3.30"
 num-traits = "0.2.16"
+rand = "0.8.5"
 reqwest = { version = "0.11.26", features = [
   "gzip",
   "deflate",
   "stream",
   "native-tls-vendored",
 ] }
-risc0-binfmt = "1.2.1"
-risc0-zkvm = { version = "1.2.1", features = ["prove"], default-features = false }
+risc0-binfmt = { workspace = true }
+risc0-zkvm = { workspace = true, features = ["prove"], default-features = false }
 serde = { version = "1.0.197" }
 serde_json = "1.0.104"
-solana-rpc-client = "~2.0"
-solana-rpc-client-api = "~2.0"
-solana-sdk = "~2.0"
-tokio = "1.36.0"
+solana-rpc-client = { workspace = true }
+solana-rpc-client-api = { workspace = true }
+solana-sdk = { workspace = true }
+tokio = { workspace = true }
